# http/docs/swagger.yaml
basePath: /api/v1
info:
  contact: {}
paths:
...
  /users/:userId/channels:
    get:
      consumes:
      - application/json
      description: get user channel list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get user channel list
      tags:
      - channel
    post:
      consumes:
      - application/json
      description: create channel
      parameters:
      - description: channel name
        in: query
        name: name
        required: true
        type: string
      - description: channel description
        in: query
        name: description
        type: string
      - description: channel group
        in: query
        name: group
        required: true
        type: string
      - description: creator id
        in: query
        name: creator_id
        required: true
        type: string
      - collectionFormat: csv
        description: subscribe user id
        in: query
        items:
          type: string
        name: subscribe_user_id
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: create channel
      tags:
      - channel
...
swagger: "2.0"